Factor out Filter as React component
