+++ src/Transformation.tsx
@@ -1,41 +1,9 @@
 /* eslint use-isnan: 0 */
 import React, { ReactElement } from "react";
-import { useState, useEffect, useCallback } from "react";
+import { useState } from "react";
 import "./Transformation.css";
 import Error from "./Error";
-import {
-  getAllDataContexts,
-  getDataFromContext,
-  setContextItems,
-  addNewContextListener,
-  removeNewContextListener,
-  addContextUpdateListener,
-  removeContextUpdateListener,
-  createTableWithData,
-  getDataContext,
-} from "./utils/codapPhone";
-import { Value } from "./language/ast";
-import { Env } from "./language/interpret";
-import { evaluate } from "./language";
-import { CodapIdentifyingInfo } from "./utils/codapPhone/types";
-import { filter } from "./transformations/filter";
-
-function useDataContexts() {
-  const [dataContexts, setDataContexts] = useState<CodapIdentifyingInfo[]>([]);
-
-  async function refreshTables() {
-    setDataContexts(await getAllDataContexts());
-  }
-
-  // Initial refresh to set up connection, then start listening
-  useEffect(() => {
-    refreshTables();
-    addNewContextListener(refreshTables);
-    return () => removeNewContextListener(refreshTables);
-  }, []);
-
-  return dataContexts;
-}
+import { Filter } from "./transformation-components/Filter";
 
 /**
  * Transformation represents an instance of the plugin, which applies a
@@ -52,101 +20,21 @@ function Transformation(): ReactElement {
 
   const transformTypes = [TransformType.Filter];
 
-  const [inputDataCtxt, setInputDataCtxt] = useState<string | null>(null);
   const [transformType, setTransformType] =
     useState<TransformType | null>(null);
-  const [transformPgrm, setTransformPgrm] = useState("");
   const [errMsg, setErrMsg] = useState<string | null>(null);
-  const dataContexts = useDataContexts();
-  const [lastContextName, setLastContextName] = useState<string | null>(null);
 
-  function inputChange(event: React.ChangeEvent<HTMLSelectElement>) {
-    setInputDataCtxt(event.target.value);
-    setErrMsg(null);
-  }
+  const transformComponents = {
+    Filter: <Filter setErrMsg={setErrMsg} />,
+  };
 
   function typeChange(event: React.ChangeEvent<HTMLSelectElement>) {
     setTransformType(event.target.value as TransformType);
     setErrMsg(null);
   }
 
-  function pgrmChange(event: React.ChangeEvent<HTMLTextAreaElement>) {
-    setTransformPgrm(event.target.value);
-    setErrMsg(null);
-  }
-
-  /**
-   * Applies the user-defined transformation to the indicated input data,
-   * and generates an output table into CODAP containing the transformed data.
-   */
-  const transform = useCallback(
-    async (doUpdate: boolean) => {
-      if (inputDataCtxt === null) {
-        setErrMsg("Please choose a valid data context to transform.");
-        return;
-      }
-      if (transformType === null) {
-        setErrMsg("Please choose a valid transformation type.");
-        return;
-      }
-
-      console.log(`Data context to transform: ${inputDataCtxt}`);
-      console.log(`Transformation type: ${transformType}`);
-      console.log(`Transformation to apply:\n${transformPgrm}`);
-
-      const dataset = {
-        context: await getDataContext(inputDataCtxt),
-        records: await getDataFromContext(inputDataCtxt),
-      };
-
-      try {
-        const filtered = filter(dataset, { predicate: transformPgrm });
-
-        // if doUpdate is true then we should update a previously created table
-        // rather than creating a new one
-        if (doUpdate) {
-          if (!lastContextName) {
-            setErrMsg("Please apply transformation to a new table first.");
-            return;
-          }
-          setContextItems(lastContextName, filtered.records);
-        } else {
-          const [newContext, _newTable] = await createTableWithData(
-            inputDataCtxt,
-            filtered.records
-          );
-          setLastContextName(newContext.name);
-        }
-      } catch (e) {
-        setErrMsg(e.message);
-      }
-    },
-    [inputDataCtxt, transformType, transformPgrm, lastContextName]
-  );
-
-  useEffect(() => {
-    if (inputDataCtxt !== null) {
-      addContextUpdateListener(inputDataCtxt, () => {
-        transform(true);
-      });
-      return () => removeContextUpdateListener(inputDataCtxt);
-    }
-  }, [transform, inputDataCtxt]);
-
   return (
     <div className="Transformation">
-      <p>Table to Transform</p>
-      <select id="inputDataContext" onChange={inputChange}>
-        <option selected disabled>
-          Select a Data Context
-        </option>
-        {dataContexts.map((dataContext) => (
-          <option key={dataContext.name} value={dataContext.name}>
-            {dataContext.title} ({dataContext.name})
-          </option>
-        ))}
-      </select>
-
       <p>Transformation Type</p>
       <select id="transformType" onChange={typeChange}>
         <option selected disabled>
@@ -156,17 +44,7 @@ function Transformation(): ReactElement {
           <option key={i}>{type}</option>
         ))}
       </select>
-
-      <p>How to Transform</p>
-      <textarea onChange={pgrmChange}></textarea>
-
-      <br />
-      <button onClick={() => transform(false)}>
-        Create table with transformation
-      </button>
-      <button onClick={() => transform(true)} disabled={!lastContextName}>
-        Update previous table with transformation
-      </button>
+      {transformType && transformComponents[transformType]}
 
       <Error message={errMsg} />
     </div>
